fix: 로그인 여부에 따른 푸터 표시 내용 수정
